requirements.txt
telethon>=1.36.0
httpx>=0.27.0
tenacity>=8.2.3
python-dotenv>=1.0.1
APScheduler>=3.10.4
orjson>=3.10.7
rich>=13.7.1
